Fix clang warnings

+++ solver/include/solver/solvercontext.hpp
@@ -377,7 +377,7 @@ public:
    * ### remarks  Aae, 30.09.2018.
    */
 
-  size_t own_conf_number() const;
+  uint8_t own_conf_number() const;
 
   /**
    * @fn  size_t SolverContext::cnt_trusted() const;
