+++ solver/src/states/trustedstage1state.cpp
@@ -19,7 +19,7 @@ void TrustedStage1State::on(SolverContext& context) {
 
   DefaultStateBehavior::on(context);
   context.init_zero(stage);
-  stage.sender = static_cast<uint8_t>(context.own_conf_number());
+  stage.sender = context.own_conf_number();
 
   enough_hashes = false;
   transactions_checked = false;
